refactor(renderers): split keycap metrics into per-style files
