+++ Apps/Keyty/Sources/Keyty/Features/Visualizers/Keyboard/Renderers/Apple/AppleKeycapMetrics.swift
@@ -1,0 +1,17 @@
+//
+//  AppleKeycapMetrics.swift
+//  Keyty
+//
+//  SPDX-FileCopyrightText: 2026 Serhii Bykov
+//  SPDX-License-Identifier: BSD-3-Clause
+//
+
+import AppKit
+
+enum AppleKeycapMetrics {
+    static let height: CGFloat = 74
+    static let minWidth: CGFloat = 74
+    static let itemSpacing: CGFloat = 6
+    static let groupPadding = NSEdgeInsets(top: 10, left: 10, bottom: 18, right: 10)
+    static let repeatBadgeInset = CGPoint(x: 6, y: 6)
+}
