+++ Apps/Keyty/Sources/Keyty/Features/Visualizers/Keyboard/Renderers/Apple/AppleKeycapRenderer.swift
