+++ Apps/Keyty/Sources/Keyty/Features/Visualizers/Keyboard/Renderers/Minimal/MinimalKeycapRenderer.swift
