+++ Apps/Keyty/Sources/Keyty/Features/Visualizers/Keyboard/Renderers/Retro/RetroKeycapMetrics.swift
@@ -1,0 +1,25 @@
+//
+//  RetroKeycapMetrics.swift
+//  Keyty
+//
+//  SPDX-FileCopyrightText: 2026 Serhii Bykov
+//  SPDX-License-Identifier: BSD-3-Clause
+//
+
+import AppKit
+
+enum RetroKeycapMetrics {
+    static let height: CGFloat = 96
+    static let bodyCornerRadius: CGFloat = 24
+    static let faceCornerRadius: CGFloat = 20
+    static let faceSideInset: CGFloat = 4
+    static let extraWidth: CGFloat = 8
+    static let faceTopInset: CGFloat = 7
+    static let faceBottomInset: CGFloat = 14
+    static let minWidth: CGFloat = 96
+    static let horizontalPadding: CGFloat = 24
+    static let itemSpacing: CGFloat = 10
+    static let groupPadding = NSEdgeInsets(top: 12, left: 12, bottom: 16, right: 12)
+    static let repeatBadgeInset = CGPoint(x: 6, y: 6)
+    static let pressedTravel: CGFloat = 3
+}
